+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   `PageInterface`
 * fix(`QueryBuilderAdapter`): throws an exception if the query does not have an
   order by clause.
+* feat: add `PagerfantaPageable`
 
 ## 0.5.0